refactor: Rename route object file
