+++ mission/src/router/index.js
@@ -1,5 +1,5 @@
 import { createRouter, createWebHistory } from 'vue-router';
-import routes from './routes';
+import routes from './RoutesObj';
 
 const router = createRouter({
   history: createWebHistory(process.env.BASE_URL),
